Serialize code complete candidates into a file.
We again stumbled upon a problem where limit on IPC argument length is
imposed by the OS. E.g. when we try to code-complete on 'std::' we get
more than 2000 completion candidates which breaks this limit on several
PCs I've tested on.

As serializing to a file seems a little bit awkard, next step would be
to figure out how to use Vim channels or jobs to implement this feature.