Prefer 'cxxd#utils#is_more_modifications_done' over '&modified'.
'&modified' flag is not enough to detect when we are undoing the
text-change in which case flag is not going be set. If we fail to detect
such cases then we are going to miss sending an updated version of contents
file and we will therefore get mismatched results from cxxd server with respect
to lines and/or columns. E.g. semantic-syntax-highlighting applied to wrong
lines and/or rows. 'cxxd#utils#is_more_modifications_done()' is able to
detect changes in both 'directions' and therefore suits very well for this
use-case.